application/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-vision-computervision==0.9.0
black==22.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.95.0
filelock==3.10.2
flake8==6.0.0
frozenlist==1.3.3
h11==0.14.0
huggingface-hub==0.13.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pytest==7.2.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
regex==2023.3.22
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.8.2